Move code around so that Database subclasses can overwrite behaviour
