+++ requirements-dev.txt
@@ -22,4 +22,4 @@ sphinx==2.2.0
 sphinx-rtd-theme==0.4.3
 
 # Formatting
-black==20.8b1
+black==21.5b1
